Move UI logic to separate folder.
